apps/langextract/pyproject.toml
[project]
name = "j16z-langextract"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langextract==1.1.1",
    "bullmq>=2.19.5",
    "redis>=5.0.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.1.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-regressions>=0.6.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
